Ajustes de diseño

+++ app/src/main/java/ar/edu/ort/respirar/data/dto/station/Precipitation.kt
@@ -2,5 +2,5 @@ package ar.edu.ort.respirar.data.dto.station
 
 data class Precipitation(
     var type: String?,
-    var value: Int?
+    var value: Double?
 )
